requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiosignal==1.4.0
anyio==4.12.0
attrs==25.4.0
certifi==2025.11.12
charset-normalizer==3.4.4
frozenlist==1.8.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
multidict==6.7.0
nodejs-wheel-binaries==24.12.0
propcache==0.4.1
python-dotenv==1.2.1
pytubefix==10.3.6
requests==2.32.5
typing_extensions==4.15.0
urllib3==2.6.2
yarl==1.22.0
